Don't include Client keys in Provider
They're only needed during dispatch, so we can supply them when
constructing the Plugin (once), and make Provider definition (frequent)
lighter weight.

This did require making the pAuthorizeEndpoint a functional field to
support Nylas which includes the client_id int he authorization
endpoint.

It may be the case that extra parameters are fine in the authorization
endpoint, in which case I would prefer to include client_id for all
Providers within the dispatch (like we do state) and go back to a simple
field. But that would require testing all the providers, which would be
nice, but is just not feasible at this time.