Simplify Provider interface further
If we assume JSON response bodies (all are currently), we can pull the
either-decode out of all the definition sites. If we add a ToIdent
class, we can use the same trick there. We keep the phantom type a in
the Provider because that's how we specify the FromJSON and ToIdent
behavior, through the type. And GeneralizedNewtypeDeriving for the
ToIdent instances makes things even slimmer.